% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convert_to_unicode}
\alias{.convert_to_unicode}
\title{Convert \code{text} to Unicode}
\usage{
.convert_to_unicode(text)
}
\arguments{
\item{text}{character scalar to convert to unicode}
}
\value{
input text, converted to unicode if applicable
}
\description{
See documentation for \code{Encoding} for more information.
Assumes utf-8 input.
}
\keyword{internal}
