% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{.infer_case_from_vocab}
\alias{.infer_case_from_vocab}
\title{Determine Vocabulary Casedness}
\usage{
.infer_case_from_vocab(vocab)
}
\arguments{
\item{vocab}{The vocabulary as a character vector.}
}
\value{
TRUE if the vocabulary is cased, FALSE if uncased.
}
\description{
Determine whether or not a wordpiece vocabulary is case-sensitive.
}
\details{
If none of the tokens in the vocabulary start with a capital letter, it will
be assumed to be uncased. Note that tokens like "\[CLS\]" contain uppercase
letters, but don't start with uppercase letters.
}
\keyword{internal}
