% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_chinese_char_cp}
\alias{.is_chinese_char_cp}
\title{Check whether cp is the codepoint of a CJK character.}
\usage{
.is_chinese_char_cp(cp)
}
\arguments{
\item{cp}{A unicode codepoint, as an integer.}
}
\value{
Logical TRUE if cp is codepoint of a CJK character.
}
\description{
Check whether cp is the codepoint of a CJK character.
}
\keyword{internal}
