% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_control}
\alias{.is_control}
\title{Check whether \code{char} is a control character.}
\usage{
.is_control(char)
}
\arguments{
\item{char}{A character scalar, comprising a single unicode character.}
}
\value{
TRUE if \code{char} is a control character.
}
\description{
"\\t", "\\n", and "\\r" are technically control characters but we treat them
as whitespace since they are generally considered as such.
}
\keyword{internal}
