% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{.load_or_retrieve_file}
\alias{.load_or_retrieve_file}
\title{Load a vocab or lookup file, or retrieve from cache}
\usage{
.load_or_retrieve_file(
  file,
  load_function,
  use_cache = TRUE,
  cache_dir = get_cache_dir()
)
}
\arguments{
\item{file}{Character; path to file to load.}

\item{load_function}{Function to call to load vocabuary or lookup from txt.}

\item{use_cache}{Logical; if TRUE, will attempt to retrieve the vocab or
lookup table from the specified cache location, or, if not found there,
will ask to save the vocabulary or lookup as an .rds file.}

\item{cache_dir}{Character; the path to a cache directory (defaults to
location returned by \code{get_cache_dir()}).}
}
\value{
The lookup table or vocabulary as returned by \code{load_function}.
}
\description{
Load a vocab or lookup file, or retrieve from cache
}
\keyword{internal}
