% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_cache_filename}
\alias{.make_cache_filename}
\title{Construct Cache File Name}
\usage{
.make_cache_filename(vocab_file)
}
\arguments{
\item{vocab_file}{path to vocabulary file. File is assumed to be a text file,
with one token per line, with the line number (starting at zero)
corresponding to the index of that token in the vocabulary.}
}
\value{
A unique filename to use for cacheing the vocabulary.
}
\description{
Given the path to a vocabulary file, construct a unique filename using the
hash of the path.
}
\keyword{internal}
