% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{.mp_tokenize_word}
\alias{.mp_tokenize_word}
\title{Apply Morphemepiece tokenization to word}
\usage{
.mp_tokenize_word(word, vocab)
}
\arguments{
\item{word}{Character; word to tokenize.}

\item{vocab}{A morphemepiece vocabulary.}
}
\value{
A character vector corresponding to the original word, broken into
pieces.
}
\description{
Repeatedly pull out the longest valid vocabulary piece from the word until
the whole word is represented in terms of vocabulary pieces. This algorithm
is intended to be used as a "fall-back" for words that are not found in the
morphemepiece lookup table.
}
\keyword{internal}
