% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{.mp_tokenize_word_bidir}
\alias{.mp_tokenize_word_bidir}
\title{Tokenize a Word Bidirectionally}
\usage{
.mp_tokenize_word_bidir(word, vocab)
}
\arguments{
\item{word}{Word to tokenize.}

\item{vocab}{Named integer vector containing vocabulary words. Should have
"vocab_split" and "rev_vocab" attributes, with components named "prefixes",
"words", "suffixes".}
}
\value{
Input word as a list of tokens.
}
\description{
Apply .mp_tokenize_word from both directions and pick the result with fewer
pieces.
}
\keyword{internal}
