% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.new_morphemepiece_vocabulary}
\alias{.new_morphemepiece_vocabulary}
\title{Constructor for Class morphemepiece_vocabulary}
\usage{
.new_morphemepiece_vocabulary(vocab, vocab_split, rev_vocab, is_cased)
}
\arguments{
\item{vocab}{Named integer vector; the "actual" vocabulary.}

\item{vocab_split}{List of named integer vectors; the split vocabulary.}

\item{rev_vocab}{List of named integer vectors; split + reversed vocabulary.}

\item{is_cased}{Logical; whether the vocabulary is cased.}
}
\value{
The vocabulary with \code{is_cased} attached as an attribute, and the
class \code{morphemepiece_vocabulary} applied. The split and reversed
vocabularies are also attached as attributes.
}
\description{
Constructor for Class morphemepiece_vocabulary
}
\keyword{internal}
