% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{.process_mp_vocab}
\alias{.process_mp_vocab}
\alias{.process_mp_vocab.default}
\alias{.process_mp_vocab.morphemepiece_vocabulary}
\alias{.process_mp_vocab.integer}
\alias{.process_mp_vocab.character}
\title{Process a Morphemepiece Vocabulary for Tokenization}
\usage{
.process_mp_vocab(v)

\method{.process_mp_vocab}{default}(v)

\method{.process_mp_vocab}{morphemepiece_vocabulary}(v)

\method{.process_mp_vocab}{integer}(v)

\method{.process_mp_vocab}{character}(v)
}
\arguments{
\item{v}{An object of class \code{morphemepiece_vocabulary}.}
}
\value{
A character vector of tokens for tokenization.
}
\description{
Process a Morphemepiece Vocabulary for Tokenization
}
\keyword{internal}
