% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{.pull_out_longest_piece}
\alias{.pull_out_longest_piece}
\title{Pull the longest piece out of the given word}
\usage{
.pull_out_longest_piece(
  word,
  mp_vocab_nohash,
  original_start = TRUE,
  original_end = TRUE,
  prefixes
)
}
\arguments{
\item{word}{Character; word to break into pieces.}

\item{mp_vocab_nohash}{A morphemepiece vocabulary with ## in names only.}

\item{original_start}{Logical; is the beginning of this word the beginning of
the original word?}

\item{original_end}{Logical; is the end of this word the end of the original
word?}

\item{prefixes}{List of prefixes from vocab. (I still need to clean this up.)}
}
\value{
A list of three "character" elements. The middle piece is the piece
"pulled out" by this function, the first and last are the left and right
remainders, respectively. They may be empty (""). (If an empty word is input,
the output will be simply "". This is an inconsistency which we may want to
deal with.)
}
\description{
Valid pieces meet the following restrictions: all internal "breaks" have at
least one "##" between, and no external edges have a "##".
}
\keyword{internal}
