% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tokenize_chinese_chars}
\alias{.tokenize_chinese_chars}
\title{Add whitespace around any CJK character.}
\usage{
.tokenize_chinese_chars(text)
}
\arguments{
\item{text}{A character scalar.}
}
\value{
Text with spaces around CJK characters.
}
\description{
Add whitespace around any CJK character.
}
\keyword{internal}
