% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_morphemepiece_vocabulary}
\alias{.validate_morphemepiece_vocabulary}
\title{Validator for Objects of Class morphemepiece_vocabulary}
\usage{
.validate_morphemepiece_vocabulary(vocab)
}
\arguments{
\item{vocab}{morphemepiece_vocabulary object to validate}
}
\value{
\code{vocab} if the object passes the checks. Otherwise, abort with
message.
}
\description{
Validator for Objects of Class morphemepiece_vocabulary
}
\keyword{internal}
