% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.whitespace_tokenize}
\alias{.whitespace_tokenize}
\title{Run basic whitespace cleaning and splitting on a piece of text.}
\usage{
.whitespace_tokenize(text)
}
\arguments{
\item{text}{Character scalar to tokenize.}
}
\value{
Character vector of tokens, split on whitespace.
}
\description{
Run basic whitespace cleaning and splitting on a piece of text.
}
\keyword{internal}
