% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Retrieve Directory for vocabulary Cache}
\usage{
get_cache_dir()
}
\value{
A unique filename to use for cacheing the vocabulary.
}
\description{
Retrieve Directory for vocabulary Cache
}
