% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{load_or_retrieve_vocab}
\alias{load_or_retrieve_vocab}
\title{Load a vocabulary file, or retrieve from cache}
\usage{
load_or_retrieve_vocab(
  vocab_file,
  use_cache = TRUE,
  cache_dir = get_cache_dir()
)
}
\arguments{
\item{vocab_file}{path to vocabulary file. File is assumed to be a text file,
with one token per line, with the line number (starting at zero)
corresponding to the index of that token in the vocabulary.}

\item{use_cache}{Logical; if TRUE, will attempt to retrieve the vocabulary
from the specified cache location, or, if not found there, will ask to save
the vocabulary as an .rds file.}

\item{cache_dir}{Character; the path to a cache directory (defaults to
location returned by \code{get_cache_dir()}).}
}
\value{
The vocab as a list of named integer vectors. Names are tokens in
vocabulary, values are integer indices. The casedness of the vocabulary is
inferred and attached as the "is_cased" attribute.

Note that from the perspective of a neural net, the numeric indices \emph{are}
the tokens, and the mapping from token to index is fixed. If we changed the
indexing, it would break any pre-trained models. This is why the vocabulary
is stored as a named integer vector, and why it starts with index zero.
}
\description{
Load a vocabulary file, or retrieve from cache
}
\examples{
# Get path to sample file included with package.
# vocab_path <- system.file("extdata", "file.txt", package = "morphemepiece")
# vocab <- load_or_retrieve_vocab(vocab_file = vocab_path, use_cache = FALSE)
}
