% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{morphemepiece_cache_dir}
\alias{morphemepiece_cache_dir}
\title{Retrieve Directory for Morphemepiece Cache}
\usage{
morphemepiece_cache_dir()
}
\value{
A character vector with the normalized path to the cache.
}
\description{
The morphemepiece cache directory is a platform- and user-specific path where
morphemepiece saves caches (such as a downloaded lookup). You can override the
default location in a few ways:
\itemize{
\item{Option: \code{morphemepiece.dir}}{Use
\code{\link{set_morphemepiece_cache_dir}} to set a specific cache directory
for this session}
\item{Environment: \code{MORPHEMEPIECE_CACHE_DIR}}{Set this environment
variable to specify a morphemepiece cache directory for all sessions.}
\item{Environment: \code{R_USER_CACHE_DIR}}{Set this environment variable
to specify a cache directory root for all packages that use the caching
system.}
}
}
