% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{.data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:tidyeval-data]{.data}}}
}}

