% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{.new_morphemepiece_vocabulary}
\alias{.new_morphemepiece_vocabulary}
\title{Constructor for Class morphemepiece_vocabulary}
\usage{
.new_morphemepiece_vocabulary(vocab, vocab_split, is_cased)
}
\arguments{
\item{vocab}{Character vector; the "actual" vocabulary.}

\item{vocab_split}{List of character vectors; the split vocabulary.}

\item{is_cased}{Logical; whether the vocabulary is cased.}
}
\value{
The vocabulary with \code{is_cased} attached as an attribute, and the
class \code{morphemepiece_vocabulary} applied. The split vocabulary is also
attached as an attribute.
}
\description{
Constructor for Class morphemepiece_vocabulary
}
\keyword{internal}
