% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{load_lookup}
\alias{load_lookup}
\title{Load a morphemepiece lookup file}
\usage{
load_lookup(lookup_file)
}
\arguments{
\item{lookup_file}{path to vocabulary file. File is assumed to be a text
file, with one word per line. The lookup value, if different from the word,
follows the word on the same line, after a space.}
}
\value{
The lookup as a named list. Names are words in lookup.
}
\description{
Load a morphemepiece lookup file
}
\examples{
# {morphemepiece} comes with a lookup that can be accessed via
# `morphemepiece_lookup()`. 
# This function can be used to load a different lookup from a file.
\dontrun{
vocab <- load_lookup("path/to/file/lookup.txt")
}
}
