% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{load_or_retrieve_lookup}
\alias{load_or_retrieve_lookup}
\title{Load a lookup file, or retrieve from cache}
\usage{
load_or_retrieve_lookup(lookup_file)
}
\arguments{
\item{lookup_file}{path to vocabulary file. File is assumed to be a text
file, with one word per line. The lookup value, if different from the word,
follows the word on the same line, after a space.}
}
\value{
The lookup table as a named character vector.
}
\description{
Load a lookup file, or retrieve from cache
}
\examples{
# {morphemepiece} comes with a lookup that can be accessed via
# `morphemepiece_lookup()`. 
# This function can be used to load (and cache) a different lookup from a 
# file.
\dontrun{
vocab <- load_or_retrieve_lookup("path/to/file/lookup.txt")
}
}
