% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{load_or_retrieve_lookup}
\alias{load_or_retrieve_lookup}
\title{Load a lookup file, or retrieve from cache}
\usage{
load_or_retrieve_lookup(
  lookup_file,
  use_cache = TRUE,
  cache_dir = get_cache_dir()
)
}
\arguments{
\item{lookup_file}{path to vocabulary file. File is assumed to be a text file,
with one word per line. The lookup value, if different from the word,
follows the word on the same line, after a space.}

\item{use_cache}{Logical; if TRUE, will attempt to retrieve the lookup
from the specified cache location, or, if not found there, will ask to save
the lookup as an .rds file.}

\item{cache_dir}{Character; the path to a cache directory (defaults to
location returned by \code{get_cache_dir()}).}
}
\value{
The lookup table as a named character vector.
}
\description{
Load a lookup file, or retrieve from cache
}
\examples{
# Get path to sample vocabulary included with package.
# vocab_path <- system.file("extdata", "tiny_vocab.txt", package = "wordpiece")
# vocab <- load_or_retrieve_vocab(vocab_file = vocab_path, use_cache = FALSE)
}
