% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphemepiece.R
\name{morphemepiece_tokenize}
\alias{morphemepiece_tokenize}
\title{Tokenize Sequence with Morpheme Pieces}
\usage{
morphemepiece_tokenize(text, vocab)
}
\arguments{
\item{text}{Character scalar; text to tokenize.}

\item{vocab}{A morphemepiece vocabulary.}
}
\value{
A character vector of tokenized text (later, this should be a named
integer vector, as in the wordpiece package.)
}
\description{
Given a single sequence of text and a morphemepiece vocabulary, tokenizes the
text. (This is \emph{not} fully functional yet. It's mainly a placeholder so that
the package exports something.)
}
