% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{.data}
\alias{\%||\%}
\alias{\%fin\%}
\alias{morphemepiece_vocab}
\alias{morphemepiece_lookup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fastmatch}{\code{\link[fastmatch:fmatch]{\%fin\%}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{morphemepiece.data}{\code{\link[morphemepiece.data]{morphemepiece_lookup}}, \code{\link[morphemepiece.data]{morphemepiece_vocab}}}

  \item{rlang}{\code{\link[rlang:op-null-default]{\%||\%}}, \code{\link[rlang:dot-data]{.data}}}
}}

